re-license under lgplv3+apache
In fc8e4c5fa19d11afb4d9ebed8a94082b7d9459ae, the intention was to
"relicenses heketi" with the rule that "rest api client code is under
dual apache2 or lgpl3+"

However some files under pkg/utils that are needed by the rest api
client was wrongly marked as "dual lgpl3+ or gpl2".

After that there was an effort to move code that's not needed by rest
api out of pkg/utils and now we just have 3 go files and the related
test code which are all needed by the rest api.

So can we please switch them back over to "dual apache2 or lgpl3+" so we
can use these in kubernetes.

Change-Id: I657621a91bc767b101781e27630c989e91d28bd7